Feat: 댓글 및 게시글 삭제 로직 추가 및 삭제 플래그(isDeleted) 관련 기능 구현

See merge request s12-final/S12P31E105!172

+++ backend/ExampleService/src/main/java/com/barcoder/scrud/post/application/service/PostService.java
@@ -137,6 +137,11 @@ public class PostService {
         Post post = postJpaRepository.findById(postId)
                 .orElseThrow(() -> new BaseException(PostErrorStatus.POST_NOT_FOUND));
 
+        // 게시글 작성자와 요청자가 다를 경우 예외 처리
+        if (!post.getUserId().equals(userId)) {
+            throw new BaseException(PostErrorStatus.POST_NOT_AUTHORIZED);
+        }
+
         // 게시글 삭제
         postJpaRepository.delete(post);
     }
